requirements.txt
# Core libraries
ccxt==3.0.36
polars==0.16.8
pandas==1.5.3
numpy==1.23.5        # Stabilna wersja kompatybilna z Python 3.11
requests==2.32.3     # Najnowsza wersja kompatybilna
PyYAML==6.0
joblib==1.3.1

# Machine Learning
scikit-learn==1.5.0
matplotlib~=3.10.0
# Logging and utilities
rich==13.5.2

aiohttp~=3.11.11

# Deep Learning
tensorflow==2.12.1  # Stabilna wersja dla Twojego Pythona

aiohttp~=3.11.11
aiofiles